DeepLearning_with_Docker/huggingface/requirements.txt
# requirements.txt
torch
torchvision
Pillow
gradio
